Added Abbaration Channel Options
Used previous functions created to create a function the applies abbaratiosn corresponding to the Zernike polynomial indices onto the incoming beam. Current code added an array of strenths to the abbaration channel, assuming that each Zerinke polynomial can be applied with different strength. Note that this code was based on current understanding of how Nazanin's code works, but further consultation may be required to ensure that it is applied properly. Nazanin's code seems to use a normalized HyGG beam with units normalized to the beam waist, and an apperature which is automatically set to 10*weist, while current code simply allows the user to choose the apperature of the abbaration channel.